Db optimizations dapper with migrations (#371)
* Semi-working skill and PC skill caching.

* Migrate most of the skill service db calls to use the cache. Update DB calls to be async.

* Remove the DB async calls. Bad idea.

* Start work on the database background thread.

* Add methods to retrieve data using dapper. Remove EF references from the skill service. Set up DB background thread to process requests and output errors if any pop up.

* Remove the EF data context but leave in the contract until I get ready to do DB call changes. Server app will not run until this is all migrated.

* Remove EF dependencies from the data editor tool. Add object builders for some of the common entities.

* Change namespace for entities. Fix up some issues with caching. Lock the data service down to only work with ICacheable objects instead of all entities. Add a way to run the object builder procs against a single record.

* Starting the individual DB call migration process.

* More DB call updates

* More DB fixes.

* Get the code compiling again.

* Remove navigation properties / child objects from all entity objects to reduce caching complexity.

* Fix db creation and migrations.

* Cache everything to get running. Add brackets around table names. Server runs again but throwing DB issues due to Dapper's inability to determine type at runtime.

* Remove Dapper.Contrib. Move that library into the solution source code. Add a way to insert/update/delete based on a runtime type.

* Rewrite a few stored procs into LINQ queries against the cached data.

* Fixed some loading crashes, cleaned out some unnecessary code. Fixed a couple Area queries.

* Remove SQL proc calls from perk and skill services.

* Fix a few bugs in the skills and perk services.

* Drop unused procs script.

* Don't apply stat changes twice during initialization.

* Begin migration from ints/string primary keys to guid/uniqueidentifiers

* Switch app-generated IDs

* Update schema script for fixed ID types

* Singularize database entities. Clean up their names.

* Add a script for generating the DB with guid primary keys.

* Auto gen IDs on entity constructor. Remove calls to shortened form of guids.

* Fix some columns which weren't nullable but should have been. Refactor the residency bonus check to fix some bugs and make it cleaner to read.

* Replace the initialization script with the latest schema. Remove the temp scripts and update the WipePlayerData script. Fix a null check on the farming service. Remove debug messages from the loot service.

* Only check for crafter player guid if string is not empty.

* Switch long types to int, to work with the new DB schema. Fix the initialization script.

* Fix a null check on the blueprint header

* Null check on the item creation portion of crafting.

* Grant +1 searches if player has the scavenger background

* Fix some dialogue problems.

* Change the placeable used for supplies caches

* Add skill tagging to Force Aura and Force Push. Add enmity rule to Force Push.

+++ SWLOR.Game.Server/Service/Contracts/ICustomEffectService.cs
@@ -1,4 +1,5 @@
 ﻿using SWLOR.Game.Server.Data;
+using SWLOR.Game.Server.Data.Entity;
 using SWLOR.Game.Server.Enumeration;
 using SWLOR.Game.Server.GameObject;
 
@@ -15,7 +16,7 @@ namespace SWLOR.Game.Server.Service.Contracts
         void OnModuleLoad();
         void OnModuleEnter();
         void RemovePCCustomEffect(NWPlayer oPC, CustomEffectType customEffectType);
-        void RemovePCCustomEffect(NWPlayer oPC, long customEffectID);
+        void RemovePCCustomEffect(NWPlayer oPC, int customEffectID);
         int GetCustomEffectLevel(NWCreature creature, CustomEffectType customEffectType);
         void ApplyStance(NWPlayer player, CustomEffectType customEffect, PerkType perkType, int effectiveLevel, string data);
         bool RemoveStance(NWPlayer player, PCCustomEffect stanceEffect = null, bool sendMessage = true);
