+++ SWLOR.Game.Server/ValueObject/CraftingData.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using SWLOR.Game.Server.Data;
+using SWLOR.Game.Server.Data.Entity;
 using SWLOR.Game.Server.Enumeration;
 using SWLOR.Game.Server.GameObject;
 
@@ -7,7 +8,7 @@ namespace SWLOR.Game.Server.ValueObject
 {
     public class CraftingData
     {
-        public long BlueprintID { get; set; }
+        public int BlueprintID { get; set; }
         public CraftBlueprint Blueprint { get; set; }
         public int MainMinimum { get; set; }
         public int MainMaximum { get; set; }
